revert: fix: avoid using localstorage polyfill to support SSR
The actual issue is that most libraries need a way to identify the
server-side environment and a lot of libraries do this by checking
Node.js specific global or environment variables (like `process`). Instead of
maintaining a custom implementation for `lib0/broadcastchannel`, we
should fix the issue by providing the correct environment for libraries
that expact Node.js compatiblity via a Node.js compatiblity mode.